Add support for 'TestEnableUnstableEras' node config entry
And use it for the Mary -> Alonzo hard fork.

During development and integration of new eras we wish to be able to
test the hard fork transition into the new era, but we do not wish to
generally have the node advertise that it understands the new era.

Avoiding advertising new development eras until they are ready makes it
practical to include new not-yet-ready eras into the main release
version of the node without the danger that operators on the mainnet
will prematurely advertise that their nodes are capable of crossing the
next hard fork.

This flag should be set to true for nodes in testnets where the testnet
purpose is to test the new era. It should always remain at the default
of false for nodes on mainnet.